requirements.txt
youtube_transcript_api

google-generativeai
python-dotenv
pathlib
streamlit



PyPDF2

langchain_google_genai

pdfplumber



pdf2image
Pillow

langchain

chromadb









